Begin implementing ‘add’ command by finding the user’s editor
This patch teaches the program to accept the ‘add’ command.  However,
all the command does is try to find the user’s editor and then print
that to screen for debugging purposes.  First we look for the
environment variable $EDITOR.  If that does not exist then we run

    $ git config --get core.editor

to get the editor name from the user’s Git configuration.  And if that
produces nothing then we return an empty string, which we will use
later to signal an error.