Open the user’s editor on a temporary file to write notes
We will use a temporary file to store the notes a user wants to save
for the current branch.  The File::Temp module is standard with Perl
so this approach should be cross-platform, as we make no explicit
assumptions about the location of the temporary file.  Then to make
sure things have worked properly we read in the entire contents of the
temporary file and print them, along with the current branch name, to
standard output.