Save branch notes to the database
This patch completes the initial implementation of the ‘add’ command.
We store the notes in the database, replacing any existing ones if
they are already there.  Then we tell the user that we have saved
notes for the current branch and exit.

A glaring limitation is that we cannot edit existing notes, we would
be more useful default behavior.  We will address this in future
patches.