Test $command against an array of valid program commands
Instead of adding another `and` statement to test for the ‘rm’
command, this patch introduces an array of commands the program
recognizes as valid.  Then we can change the if-statement to an
unless-statement that grep’s that list to determine if the given
command was valid.  Now when we want to add new valid commands all we
have to do is add them to the @valid_commands array.