Edit existing notes when using the ‘add’ command
This patch teachs the program to first check for existing notes about
the current branch when using the ‘add’ command.  If any notes exist
we load those into the temporary file that the user can then edit.
This makes it easier for users to keep large amounts of notes and
periodically modify them, instead of having to start with a blank set
of notes every time.