Fix a bug where ‘add’ could lose existing notes
Running `git branch-notes add` on a branch with existing notes opens
the editor with those notes.  However, if we immediately save and exit
the program will not save the existing notes.  This is because after
calling load_notes_for_branch() our file-handle is not ready to read
from the beginning of the file, and thus everything before that point
is not saved.  This patch fixes that by rewinding the handle to the
beginning of the file before we read the contents of the notes file.