Do not require the common::sense module
The program required the common::sense module, which does not come
with Perl.  This patch removes that module and lifts the requirement
because there is no real gain in forcing that module on the user for
this program.

This patch also bumps the minor version number to reflect the removal
of that module.